+++ .idea/misc.xml
@@ -13,6 +13,8 @@
         <entry key="..\:/Users/tarad/AndroidStudioProjects/AllYourFavouriteFilms3/app/src/main/res/layout/activity_details.xml" value="0.15353260869565216" />
         <entry key="..\:/Users/tarad/AndroidStudioProjects/AllYourFavouriteFilms3/app/src/main/res/layout/activity_main.xml" value="0.17300724637681159" />
         <entry key="..\:/Users/tarad/AndroidStudioProjects/AllYourFavouriteFilms3/app/src/main/res/layout/film_item.xml" value="0.16032608695652173" />
+        <entry key="..\:/Users/tarad/AndroidStudioProjects/AllYourFavouriteFilms3/app/src/main/res/layout/fragment_details.xml" value="0.1852355072463768" />
+        <entry key="..\:/Users/tarad/AndroidStudioProjects/AllYourFavouriteFilms3/app/src/main/res/layout/fragment_home.xml" value="0.1911231884057971" />
         <entry key="..\:/Users/tarad/AndroidStudioProjects/AllYourFavouriteFilms3/app/src/main/res/menu/navigation.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/tarad/AndroidStudioProjects/AllYourFavouriteFilms3/app/src/main/res/menu/settings.xml" value="0.25416666666666665" />
       </map>
