+++ app/src/main/java/com/example/allyourfavouritefilms/DetailsFragment.kt
@@ -1,0 +1,37 @@
+package com.example.allyourfavouritefilms
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import kotlinx.android.synthetic.main.fragment_details.*
+
+class DetailsFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_details, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        setFilmsDetails()
+    }
+
+    private fun setFilmsDetails() {
+        //Получаем наш фильм из переданного бандла
+        val film = arguments?.get("film") as Film
+
+        //Устанавливаем заголовок
+        details_toolbar.title = film.title
+        //Устанавливаем картинку
+        details_poster.setImageResource(film.poster)
+        //Устанавливаем описание
+        details_description.text = film.description
+    }
+}
