+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,6 @@
     <string name="button5">Прочее</string>
     <string name="menu_settings_title">Settings</string>
     <string name="page_title">Search it!</string>
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="disclaimer">Постеры предоставлены https://www.movieposters.com</string>
 </resources>
